backend/requirements.txt
google-cloud-storage
ngt
